id int to long

+++ pokemon-app/src/main/java/com/example/springbootdocker2/repository/CardRepository.java
@@ -3,6 +3,6 @@ package com.example.springbootdocker2.repository;
 import com.example.springbootdocker2.pokecards.PokemonCard;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CardRepository extends JpaRepository<PokemonCard,Integer> {
+public interface CardRepository extends JpaRepository<PokemonCard,Long> {
 
 }
